schedule homepage, payment methods
