+++ pages/_app.tsx
@@ -9,7 +9,6 @@ function App({ Component, pageProps }: AppProps): ReactElement {
   return (
     <ChakraProvider>
       <Component {...pageProps} />
-      <Footer></Footer>
     </ChakraProvider>
   );
 }
