+++ pages/timetable.tsx
@@ -6,35 +6,14 @@ import {
   Stack,
   List,
   ListItem,
-  Button,
-  useColorModeValue,
   Icon,
   Heading,
   Link,
-  Spacer,
   Flex,
-  VStack,
-  ListIcon,
   Divider,
 } from "@chakra-ui/react";
 
-import { FaHome } from "react-icons/fa";
-import { MdPlace } from "react-icons/md";
 import { FiArrowLeftCircle } from "react-icons/fi";
-
-interface Services {
-  service: string;
-
-  time: string;
-}
-
-interface NavItem {
-  day: string;
-  place: string;
-  address: string;
-  city: string;
-  services: Array<Services>;
-}
 
 interface PriceItemProps {
   type: string;
@@ -69,8 +48,6 @@ const PriceItemTier = ({
   const discount = offer ? "(SALE)" : "";
   const FontWeight = offer ? 700 : 500;
 
-  const colorTextDark = offer ? "white" : "yellow.500";
-  const bgColorDark = offer ? "yellow.400" : "gray.300";
   return (
     <ListItem maxW={"200px"} color="whiteAlpha.500" p={2}>
       {label} ({type})
@@ -87,71 +64,9 @@ const PriceItemTier = ({
     </ListItem>
   );
 };
-const NAV_ITEMS: Array<NavItem> = [
-  {
-    day: "Tuesday",
-    city: "Kirkkonummi",
-    place: "Kirkkonummen Palvelukeskus",
-    address: "Rajakuja 3, Kirkkonummi",
-    services: [
-      {
-        service: "LatinoMix-Dance",
 
-        time: "18:00-19:00",
-      },
-    ],
-  },
-
-  {
-    day: "Thursday",
-    city: "Espoo",
-    place: "Matinkylän Pirtti",
-    address: "Keskipäivänkuja 4, Espoo",
-    services: [
-      {
-        service: "LatinoMix-Dance",
-
-        time: "18:00-19:00",
-      },
-    ],
-  },
-
-  {
-    day: "Sunday",
-    city: "Espoo",
-    place: "Matinkylän Pirtti",
-    address: "Keskipäivänkuja 4, Espoo",
-    services: [
-      {
-        service: "LatinoMix-EasyDance",
-
-        time: "16:00-17:00",
-      },
-      {
-        service: "LatinoMix-Dance",
-
-        time: "17:00-18:00",
-      },
-    ],
-  },
-];
-
-import { ReactNode } from "react";
+import AdultsSchedule from "components/AdultsSchedule";
 import KidsSchedule from "components/KidsSchedule";
-function PriceWrapper({ children }: { children: ReactNode }) {
-  return (
-    <Box
-      mb={4}
-      shadow="base"
-      borderWidth="1px"
-      alignSelf={{ base: "center", lg: "flex-start" }}
-      borderColor={"gray.200"}
-      borderRadius={"xl"}
-    >
-      {children}
-    </Box>
-  );
-}
 
 function timetable() {
   return (
@@ -201,61 +116,25 @@ function timetable() {
             ))}
           </List>
         </Box>
+        <Box
+          alignItems="center"
+          justifyContent="center"
+          w="full"
+          bg="gray.700"
+          p={3}
+        >
+          <Heading mt={8} as="h1" size="sm" color="green.500">
+            Payment Methods:
+          </Heading>
+          <List>
+            <ListItem color={"gray.400"}>Cash: on-site</ListItem>
+            <ListItem color={"orange.500"}>MobilePay: +358 44 3732360</ListItem>
+            <ListItem color={"orange.500"}>ref: LM "your_name" </ListItem>
+          </List>
+        </Box>
       </Flex>
-      <Stack
-        direction={{ base: "column", md: "row" }}
-        textAlign="center"
-        justify="center"
-        spacing={{ base: 4, lg: 10 }}
-        py={10}
-      >
-        {NAV_ITEMS.map((item) => (
-          <PriceWrapper key={item.day}>
-            <Box py={4} px={12}>
-              <Text
-                p={2}
-                color={"green.500"}
-                rounded={"full"}
-                fontSize={"sm"}
-                fontWeight={500}
-                bg={"green.50"}
-              >
-                {item.day}
-              </Text>
-              {item.services.map((serv) => (
-                <Stack
-                  key={serv.service}
-                  align={"flex-start'"}
-                  justify={"center"}
-                >
-                  <Text p={2} fontSize={"1xl"} fontWeight={600}>
-                    {serv.service}
-                  </Text>
-
-                  <Text color={"gray.500"}>{serv.time}</Text>
-                </Stack>
-              ))}
-            </Box>
-            <VStack bg={"gray.50"} py={4} borderBottomRadius={"xl"}>
-              <List spacing={3} textAlign="start" px={12}>
-                <ListItem>
-                  <ListIcon as={FaHome} color="green.500" />
-                  {item.place}
-                </ListItem>
-                <ListItem>
-                  <ListIcon as={MdPlace} color="green.500" />
-                  {item.address}
-                </ListItem>
-              </List>
-              {/*<Box w="80%" pt={7}>
-                <Button w="full" colorScheme="red" variant="outline">
-                  Join!
-                </Button>
-              </Box>*/}
-            </VStack>
-          </PriceWrapper>
-        ))}
-      </Stack>
+      {/* add Adults component*/}
+      <AdultsSchedule></AdultsSchedule>
       {/* add kids component*/}
       <Divider orientation="horizontal" />
       <KidsSchedule></KidsSchedule>
